Allowed Constant to take multiple options for a more concise Any+Const (#19)
Since a lot of Any use was with constants, and the error messages were
long and a bit confusing, this gives the Constant field the ability to
have a set of options natively as multiple positional arguments to the
constructor. It's backwards compatible with the old calling style and
error messages, though the introspection format has changed.