Fix: Nullable introspection incorrectly squashed all inner introspection
Introspection for the new `Nullable` field was incorrectly squashing inner introspection to a Python dict string. This breaks introspection. With this change, `Nullable` introspection defers to the first positional argument, but adds a `"nullable": True` key to the returned dictionary. If the returned dictionary has no description, the `Nullable` description will be used (if set).