[PATCH] Ensure optional_keys is a set, introspects to a list
When the compiler inlines a `tuple` for the `in` comparison, a small `tuple` is considerably faster than the same values in a `set` or a `list` (though that advantage eventually disappears for large iterables). However, when the same values in a `set`, a `list`, and a `tuple` are passed to something as an argument, the speed advantage is clearly in the favor of `set` in all scenarios:

```
In [1]: s1 = {'a', 'b', 'c', 'd', 'e', 'f', 'g'}

In [2]: s2 = ['a', 'b', 'c', 'd', 'e', 'f', 'g']

In [3]: s3 = ('a', 'b', 'c', 'd', 'e', 'f', 'g')

In [4]: %timeit 'd' in s1
42.5 ns ± 2.11 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [5]: %timeit 'd' in s2
78.7 ns ± 1.19 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [6]: %timeit 'd' in s3
85.3 ns ± 2.07 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [11]: %timeit 'h' in s1
39 ns ± 0.717 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [12]: %timeit 'h' in s2
147 ns ± 1.25 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [13]: %timeit 'h' in s3
155 ns ± 1.24 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

This changes ensures that `optional_keys` will always be a `set`, for best possible performance. It also ensures that the introspection converts `optional_keys` to a list, so that it can be serialized using msgpack and JSON. This performance enhancement mirrors a similar enhancement made to `Constant`: https://github.com/eventbrite/conformity/blob/2641061a21983e43a49e5353234be381b09ffad0/conformity/fields/basic.py#L38