[MINOR] Add support for machine-readable error codes
To better support clients that interpret received error codes to display helpful messages to humans, this adds support for machine-readable error codes to field validation. When not explicitly provided, the default error code is `INVALID`, which is appropriate in most cases. Missing keys in structures are `MISSING`. Extra keys in structures, and values that don't match expected constants, are `UNKNOWN`.