[MINOR] Enhance Dictionary field to permit ordered key documentation
In dictionary validation, the order of the keys isn't important (and, indeed, can't be enforced when dealing with Msgpack, JSON, and others). However, the order that errors are presented (for display purposes) and the order the keys are defined (for documentation purposes) might be important. This enhancement makes the `Dictionary` field aware of ordered dictionary inputs, returns validation errors in that order, extends with that order kept in mind, and returns key ordering documentation with the field introspection.

Side note: This change also fixes a minor bug: Classes that inherit from `@attr.s`-decorated classes but don't define a constructor must also be decorated with `@attr.s` in order for IDE inspection to work properly.