[MINOR] Implement #65: Add support for generic Settings objects
This commit implements #65. PySOA's `Settings` are being used for many things beyond just PySOA - several other projects not otherwise using PySOA are importing PySOA just to use its `Settings` feature. Based on this use case, `Settings` is an appropriate member of the Conformity domain, so this is the Conformity commit to migrate `Settings` from PySOA to Conformity. There will be a corresponding PySOA commit to adopt this new Conformity feature.