[PATCH] [SPHINX] Soft fail getting annotations when source not found
In the Sphinx Autodoc and LinkCode extensions, if `inspect` cannot get the source lines for a member, that by extension means the member doesn't have annotations, so there's no reason to fail hard in `get_annotations`, and there's also no reason to fail hard in generating a link to the source code.

Also, add missing newline in settings schema documentation.