[MINOR] Add currency code field (#85)
Add a new field to validate currency codes. This field is a special extension of `Constant` that ensures the value is a Unicode string that enforces the value meets the currency format as `'USD'`. It has one argument, `code_filter`, which if specified must be a `typing.Callable[[typing.AnyStr], bool]`.